수류탄 구현 (애니메이션 X), 적 AI 개선, 적 체력 HUD 구현, 적 사망시 래그돌화 구현

+++ Assets/Scripts/Game/class/Class/Sealed/DamageableData.cs
@@ -1,11 +1,11 @@
 
 public sealed class DamageableData
 {
-    public int healthPoint_Max { get; private set; }
+    public float healthPoint_Max { get; private set; }
 
     public float invincibleTime { get; private set; }
 
-    public DamageableData(int healthPoint_Max, float invincibleTime)
+    public DamageableData(float healthPoint_Max, float invincibleTime)
     {
         this.healthPoint_Max = healthPoint_Max;
